backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.10"
dependencies = [
    "aioredis>=2.0.1",
    "aiosmtplib==4.0.0",
    "alembic==1.15.1",
    "annotated-types==0.7.0",
    "anyio==4.7.0",
    "async-timeout==5.0.1",
    "asyncpg==0.30.0",
    "bcrypt==4.1.2",
    "boto3==1.37.23",
    "botocore==1.37.23",
    "certifi==2025.1.31",
    "charset-normalizer==3.4.1",
    "click==8.1.8",
    "dnspython==2.7.0",
    "ecdsa==0.19.0",
    "email-validator==2.2.0",
    "exceptiongroup==1.2.2",
    "fastapi==0.115.11",
    "greenlet==3.1.1",
    "h11==0.14.0",
    "httpcore==1.0.7",
    "httpx==0.28.1",
    "idna==3.10",
    "jmespath==1.0.1",
    "loguru==0.7.3",
    "mako==1.3.9",
    "markupsafe==3.0.2",
    "naked==0.1.32",
    "passlib[bcrypt]==1.7.4",
    "pyasn1==0.4.8",
    "pycryptodome==3.21.0",
    "pydantic==2.10.6",
    "pydantic-core==2.27.2",
    "pydantic-settings==2.8.1",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.0.1",
    "python-jose==3.4.0",
    "python-multipart==0.0.20",
    "pyyaml==6.0.2",
    "requests==2.32.3",
    "rsa==4.9",
    "s3transfer==0.11.4",
    "shellescape==3.8.1",
    "six==1.17.0",
    "sniffio==1.3.1",
    "sqlalchemy==2.0.38",
    "starlette==0.46.0",
    "typing-extensions==4.12.2",
    "urllib3==2.3.0",
    "uvicorn==0.34.0",
]
